# docs/openapi.release-gate.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PreFlight Release Gate API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.usepreflight.xyz"
    }
  ],
  "paths": {
...
    "/api/v1/resolve": {
      "post": {
        "responses": {
          "200": {
            "description": "Authenticated OnchainOS CLI listing resolution with field provenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResolutionV1"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/verify-receipt": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "receipt_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public cryptographic receipt verification. This verifies issuer authenticity and payload integrity, not target correctness."
          }
        }
      },
      "post": {
        "responses": {
          "200": {
            "description": "Public cryptographic receipt verification for JSON clients. This verifies issuer authenticity and payload integrity, not target correctness."
          }
        }
      }
    },
...
    "/api/v1/passport/{agent_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "Public owner-authorized scoped passport"
          },
          "404": {
            "description": "No passport issued"
          }
        }
      }
    },
...
    "/api/v1/receipts/{receipt_id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "receipt_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public signed receipt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptEnvelopeV1"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentResolutionV1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "agent_id": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "object",
            "properties": {
              "value": {
                "anyOf": [
                  {
                    "$ref": "#/$defs/__schema0"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "source": {
                "type": "string",
                "minLength": 1
              },
              "confidence": {
                "type": "string",
                "enum": [
                  "observed",
                  "inferred",
                  "unknown"
                ]
              }
            },
            "required": [
              "value",
              "source",
              "confidence"
            ],
            "additionalProperties": false
          },
          "description": {
            "type": "object",
            "properties": {
              "value": {
                "anyOf": [
                  {
                    "$ref": "#/$defs/__schema0"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "source": {
                "type": "string",
                "minLength": 1
              },
              "confidence": {
                "type": "string",
                "enum": [
                  "observed",
                  "inferred",
                  "unknown"
                ]
              }
            },
            "required": [
              "value",
              "source",
              "confidence"
            ],
            "additionalProperties": false
          },
          "category_code": {
            "type": "object",
            "properties": {
              "value": {
                "anyOf": [
                  {
                    "$ref": "#/$defs/__schema0"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "source": {
                "type": "string",
                "minLength": 1
              },
              "confidence": {
                "type": "string",
                "enum": [
                  "observed",
                  "inferred",
                  "unknown"
                ]
              }
            },
            "required": [
              "value",
              "source",
              "confidence"
            ],
            "additionalProperties": false
          },
          "status": {
            "type": "object",
            "properties": {
              "value": {
                "anyOf": [
                  {
                    "$ref": "#/$defs/__schema0"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "source": {
                "type": "string",
                "minLength": 1
              },
              "confidence": {
                "type": "string",
                "enum": [
                  "observed",
                  "inferred",
                  "unknown"
                ]
              }
            },
            "required": [
              "value",
              "source",
              "confidence"
            ],
            "additionalProperties": false
          },
          "services": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "service_id": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/__schema0"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "source": {
                      "type": "string",
                      "minLength": 1
                    },
                    "confidence": {
                      "type": "string",
                      "enum": [
                        "observed",
                        "inferred",
                        "unknown"
                      ]
                    }
                  },
                  "required": [
                    "value",
                    "source",
                    "confidence"
                  ],
                  "additionalProperties": false
                },
                "name": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/__schema0"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "source": {
                      "type": "string",
                      "minLength": 1
                    },
                    "confidence": {
                      "type": "string",
                      "enum": [
                        "observed",
                        "inferred",
                        "unknown"
                      ]
                    }
                  },
                  "required": [
                    "value",
                    "source",
                    "confidence"
                  ],
                  "additionalProperties": false
                },
                "type": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/__schema0"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "source": {
                      "type": "string",
                      "minLength": 1
                    },
                    "confidence": {
                      "type": "string",
                      "enum": [
                        "observed",
                        "inferred",
                        "unknown"
                      ]
                    }
                  },
                  "required": [
                    "value",
                    "source",
                    "confidence"
                  ],
                  "additionalProperties": false
                },
                "fee": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/__schema0"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "source": {
                      "type": "string",
                      "minLength": 1
                    },
                    "confidence": {
                      "type": "string",
                      "enum": [
                        "observed",
                        "inferred",
                        "unknown"
                      ]
                    }
                  },
                  "required": [
                    "value",
                    "source",
                    "confidence"
                  ],
                  "additionalProperties": false
                },
                "endpoint": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/__schema0"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "source": {
                      "type": "string",
                      "minLength": 1
                    },
                    "confidence": {
                      "type": "string",
                      "enum": [
                        "observed",
                        "inferred",
                        "unknown"
                      ]
                    }
                  },
                  "required": [
                    "value",
                    "source",
                    "confidence"
                  ],
                  "additionalProperties": false
                },
                "asset_contract": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "anyOf": [
                        {
                          "$ref": "#/$defs/__schema0"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "source": {
                      "type": "string",
                      "minLength": 1
                    },
                    "confidence": {
                      "type": "string",
                      "enum": [
                        "observed",
                        "inferred",
                        "unknown"
                      ]
                    }
                  },
                  "required": [
                    "value",
                    "source",
                    "confidence"
                  ],
                  "additionalProperties": false
                }
              },
              "required": [
                "service_id",
                "name",
                "type",
                "fee",
                "endpoint",
                "asset_contract"
              ],
              "additionalProperties": false
            }
          },
          "resolved_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "resolution_source": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "agent_id",
          "name",
          "description",
          "category_code",
          "status",
          "services",
          "resolved_at",
          "resolution_source"
        ],
        "additionalProperties": false,
        "$defs": {
          "__schema0": {
            "anyOf": [
              {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/$defs/__schema0"
                }
              },
              {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "$ref": "#/$defs/__schema0"
                }
              }
            ]
          }
        }
      },
      "ReceiptEnvelopeV1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "receipt_id": {
            "type": "string",
            "pattern": "^rcpt_[a-f0-9]{32}$"
          },
          "payload": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "preflight.receipt.v1"
              },
              "receipt_id": {
                "type": "string",
                "pattern": "^rcpt_[a-f0-9]{32}$"
              },
              "report_id": {
                "type": "string",
                "minLength": 16
              },
              "decision": {
                "type": "string",
                "enum": [
                  "RELEASE",
                  "BLOCK",
                  "UNKNOWN"
                ]
              },
              "manifest_hash": {
                "type": "string",
                "pattern": "^sha256:[a-f0-9]{64}$"
              },
              "snapshot_hash": {
                "type": "string",
                "pattern": "^sha256:[a-f0-9]{64}$"
              },
              "policy_version": {
                "type": "string",
                "minLength": 1
              },
              "settlement_ref": {
                "type": "string",
                "minLength": 1
              },
              "payer": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "price_usdt": {
                "type": "string",
                "pattern": "^\\d+(?:\\.\\d{1,6})?$"
              },
              "target_fingerprint": {
                "type": "string",
                "pattern": "^sha256:[a-f0-9]{64}$"
              },
              "issued_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
              },
              "key_id": {
                "type": "string",
                "minLength": 1
              },
              "chain_anchor": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "tx": {
                        "type": "string",
                        "pattern": "^0x[a-fA-F0-9]{64}$"
                      },
                      "contract": {
                        "type": "string",
                        "pattern": "^0x[a-fA-F0-9]{40}$"
                      }
                    },
                    "required": [
                      "tx",
                      "contract"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "scope": {
                "type": "object",
                "properties": {
                  "proves": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "issuer_authenticity",
                        "payload_integrity",
                        "snapshot_binding",
                        "policy_binding"
                      ]
                    }
                  },
                  "does_not_prove": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "semantic_correctness_of_delivery",
                        "future_behaviour",
                        "security_of_target",
                        "marketplace_endorsement"
                      ]
                    }
                  },
                  "policy_version": {
                    "type": "string",
                    "minLength": 1
                  },
                  "snapshot_hash": {
                    "type": "string",
                    "pattern": "^sha256:[a-f0-9]{64}$"
                  },
                  "valid_until": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                  }
                },
                "required": [
                  "proves",
                  "does_not_prove",
                  "policy_version",
                  "snapshot_hash",
                  "valid_until"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "type",
              "receipt_id",
              "report_id",
              "decision",
              "manifest_hash",
              "snapshot_hash",
              "policy_version",
              "settlement_ref",
              "payer",
              "price_usdt",
              "target_fingerprint",
              "issued_at",
              "key_id",
              "chain_anchor"
            ],
            "additionalProperties": false
          },
          "signature": {
            "type": "string",
            "minLength": 32
          },
          "signature_alg": {
            "type": "string",
            "const": "Ed25519"
          },
          "key_id": {
            "type": "string",
            "minLength": 1
          },
          "verify": {
            "type": "object",
            "properties": {
              "canonicalization": {
                "type": "string",
                "const": "preflight.canonical-json.v1"
              },
              "payload_hash": {
                "type": "string",
                "pattern": "^sha256:[a-f0-9]{64}$"
              },
              "pubkeys_url": {
                "type": "string",
                "format": "uri"
              }
            },
            "required": [
              "canonicalization",
              "payload_hash",
              "pubkeys_url"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "receipt_id",
          "payload",
          "signature",
          "signature_alg",
          "key_id",
          "verify"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "Error": {
        "description": "Typed API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorV1"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerCapability": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}